TST: consolidate read_epochs_bids tests into test_read.py
